Unify in a single script
