+++ parse_jenkins_builds.py
@@ -1,12 +1,15 @@
 #!/usr/bin/env python
 from __future__ import print_function
 from hashlib import sha1
-import os , re , sys , json
+import os , re , sys , json, datetime, time
 import xml.etree.ElementTree as ET
+from _py2with3compatibility import run_cmd
 from es_utils import send_payload,get_payload,resend_payload,get_payload_wscroll
+
 JENKINS_PREFIX="jenkins"
 try:    JENKINS_PREFIX=os.environ['JENKINS_URL'].strip("/").split("/")[-1]
 except: JENKINS_PREFIX="jenkins"
+LOCAL_JENKINS_URL = os.environ['LOCAL_JENKINS_URL']
 
 def findParametersAction(root):
   if root.tag=='parameters': return root
@@ -26,26 +29,119 @@ def getParameters(root, payload):
   else:
     for x in root: getParameters(x, payload)
 
+def get_current_time():
+    """Returns current time in milliseconds. """
+    current_time = datetime.datetime.utcnow() - datetime.datetime(1970, 1, 1)
+    current_time = round(current_time.total_seconds()*1000)
+    return current_time
+
+def display_build_info(build_id, build_payload):
+   """Display id, job name, build number and waiting time for a conrete build in queue. """
+   print("==> ", str(build_id) + " " + str(build_payload["job_name"]) + " #" + str(build_payload["queue_id"]))
+   wait_time = build_payload["wait_time"]/1000
+   print("Time in queue (minutes): ", str(wait_time/60))
+
+def update_payload_timestamp(build_id, queue):
+    """Updates timestamp for a given payload. """
+    id = build_id
+    payload = queue[id]
+    current_time = get_current_time()
+    payload['@timestamp'] = current_time
+    return id, payload
+
 query_running_builds = """{
 "query": {"bool": {"must": {"query_string": {"query": "job_status:Running AND jenkins_server:%s", "default_operator": "AND"}}}},
 "from": 0,
 "size": 10000
 }""" % JENKINS_PREFIX
 
-query_jenkins_queue = """{
+# Query job with in_queue=1
+query_inqueue1 = """{
+"query": {"bool": {"must": {"query_string": {"query": "in_queue: 1 AND start_time: 0 AND jenkins_server: %s", "default_operator": "AND"}}}},
+"from": 0,
+"size": 10000
+}""" % JENKINS_PREFIX
+
+# Query jobs with in_queue=0
+query_inqueue0 = """{
 "query": {"bool": {"must": {"query_string": {"query": "in_queue: 0 AND start_time: 0", "default_operator": "AND"}}}},
 "from": 0,
 "size": 10000
 }"""
 
+# Get jobs in queue from elastic search
+queue_index = 'cmssdt-jenkins-queue*'
+elements_inqueue = get_payload_wscroll(queue_index, query_inqueue1)
+es_queue = dict()
+if elements_inqueue:
+  if (not 'hits' in elements_inqueue) or (not 'hits' in elements_inqueue['hits']):
+    print("ERROR: ", elements_inqueue)
+  for entry in elements_inqueue['hits']['hits']:
+    es_queue[entry['_id']] = entry['_source']
+
+# Get jenkins queue and construct payload to be send to elastic search
+exit_code, queue_json = run_cmd('curl -s -H "OIDC_CLAIM_CERN_UPN: cmssdt; charset=UTF-8" "' + LOCAL_JENKINS_URL + '/queue/api/json?pretty=true"')
+queue_json = json.loads(queue_json)
+
+jenkins_queue = dict()
+for element in queue_json["items"]:
+    payload = dict()
+
+    job_name = element["task"]["name"]
+    queue_id = int(element["id"])
+    queue_time = int(element["inQueueSince"])
+    current_time = get_current_time()
+
+    payload['jenkins_server'] = JENKINS_PREFIX
+    payload["in_queue_since"] = queue_time
+    payload["queue_id"] = queue_id
+    payload["job_name"] = job_name
+    payload["in_queue"] = 1
+    payload["wait_time"] = current_time - queue_time
+    payload["start_time"] = 0
+
+    unique_id = JENKINS_PREFIX + ":/build/builds/" + job_name + "/" + str(queue_id) # Not a real path
+    id = sha1(unique_id).hexdigest()
+    jenkins_queue[id] = payload
+
+queue_index="cmssdt-jenkins-queue-"+str(int(((queue_time/86400000)+4)/7))
 queue_document = "queue-data"
-queue_content_hash = get_payload_wscroll("cmssdt-jenkins-queue*", query_jenkins_queue)
+
+# Update information in elastic search
+new_inqueue = [x for x in jenkins_queue.keys() if x not in es_queue.keys()]
+print("[INFO] Pushing new Jenkins builds in queue ...")
+for build_id in new_inqueue:
+    id, payload = update_payload_timestamp(build_id, jenkins_queue)
+    display_build_info(id, payload)
+    send_payload(queue_index,queue_document,id,json.dumps(payload))
+
+still_inqueue = [x for x in jenkins_queue.keys() if x in es_queue.keys()]
+print("[INFO] Updating waiting time for build that are still in queue ...")
+for build_id in still_inqueue:
+    id, payload = update_payload_timestamp(build_id, jenkins_queue)
+    payload["wait_time"] = current_time - payload["in_queue_since"]
+    display_build_info(id, payload)
+    send_payload(queue_index,queue_document,id,json.dumps(payload))
+
+no_inqueue = [str(y) for y in es_queue.keys() if y not in jenkins_queue.keys()]
+print("[INFO] Updating builds that are no longer in queue ...")
+for build_id in no_inqueue:
+    id, payload = update_payload_timestamp(build_id, es_queue)
+    payload['in_queue'] = 0
+    print("==> Cleaning up ", str(id) + " " + str(payload["job_name"]) + " #" + str(payload["queue_id"]))
+    send_payload(queue_index,queue_document,id,json.dumps(payload))
+
+time.sleep(10)
+
+# Get jobs in elastic search with in_queue=0 (jobs that already started)
+queue_content_hash = get_payload_wscroll("cmssdt-jenkins-queue*", query_inqueue0)
 es_queue = dict()
 for entry in queue_content_hash['hits']['hits']:
     queue_id = entry['_source']['queue_id']
     entry['_source']['queue_hash'] = entry['_id']
     es_queue[queue_id] = entry['_source']
 
+print("[INFO] Checking status of running/finished builds ...")
 all_local = []
 path = '/build/builds'
 document = "builds-data"
@@ -74,7 +170,6 @@ for root, dirs, files in os.walk(path):
           payload['slave_node'] = 'unknown'
         try:
           payload['queue_id'] = root.find('queueId').text
-          print("Queue id: ", payload['queue_id'])
         except:
           payload['queue_id'] = 'unknown'
         payload['jenkins_server'] = JENKINS_PREFIX
@@ -88,18 +183,14 @@ for root, dirs, files in os.walk(path):
           payload['job_status'] = 'Running'
 
         # Check if job has been in queue, and update queue waiting time
-        print("[DEBUG] payload queue id: ", payload['queue_id'])
-        print("[DEBUG] es queue list: ", es_queue.keys())
         if int(payload['queue_id']) in es_queue.keys():
-            print("Updating queue waiting time for ", queue_payload)
-            queue_payload = es_queue[payload['queue_id']]
+            queue_payload = es_queue[int(payload['queue_id'])]
             queue_payload['start_time'] = int(jstime) # start time in millisec
             queue_payload['wait_time'] = int(jstime) - queue_payload["in_queue_since"]
             queue_payload['build_number'] = payload['build_number']
             
-            queue_index="cmssdt-jenkins-queue-"+str(int((((int(jstime)/1000)/86400)+4)/7))
-            print("[DEBUG] queue id: ", queue_payload['_id'])
-            send_payload(queue_index, queue_document, queue_payload['_id'], json.dumps(queue_payload))
+            print("==> Sending payload for ", queue_payload['queue_hash'])
+            send_payload(queue_index, queue_document, queue_payload['queue_hash'], json.dumps(queue_payload))
 
         all_local.append(id)
         weekindex="jenkins-jobs-"+str(int((((int(jstime)/1000)/86400)+4)/7))
